Refactor redux-thunk
